fix: file name movieBoard >> movieList, *.jar 추가 그 외 수정 중
