+++ src/main/webapp/WEB-INF/command.properties
@@ -38,7 +38,7 @@
 /admin/movieWriteDB.do=movie.service.MovieWriteDBService
 
 # 영화 목록
-/admin/movieBoard.do=movie.service.MovieBoardService
+/admin/movieList.do=movie.service.MovieListService
 
 # 영화 조회
 /admin/movieView.do=movie.service.MovieViewService
